+++ CHANGELOG.md
@@ -27,6 +27,9 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
  - Deprecated the [configuration options shown above](https://commonmark.thephpleague.com/1.6/upgrading/)
  - Deprecated the ability to pass a custom `Environment` into the constructors of `CommonMarkConverter` and `GithubFlavoredMarkdownConverter`; use `MarkdownConverter` instead
  - Deprecated `ConfigurableEnvironmentInterface::setConfig()`; use `mergeConfig()` instead
+ - Deprecated calling `ConfigurableEnvironmentInterface::mergeConfig()` without any parameters
+ - Deprecated calling `Configuration::get()` without any parameters
+ - Deprecated calling `Configuration::set()` without the second `$value` parameter
  - Deprecated `RegexHelper::matchAll()`; use `RegexHelper::matchFirst()` instead
 
 ### Fixed